Fix all react-doctor findings and add a CI scan
Full react-doctor scan scored 73/100 with 33 warnings. This fixes or
suppresses every finding, bringing the repo to 100/100:

- Remove nine unused dependencies and four unused shadcn leftover files
- Drop unused exports (useThemeContext, supportedDatabases, buttonVariants)
- Replace useContext with React 19 use()
- Move TitleBar window handlers to module scope
- Add the missing databaseId dependency to the connection test callback
- Version the theme localStorage key (theme:v1) with a legacy fallback
- Run independent awaits concurrently with Promise.all in collections
  handlers and the connection info migration
- Lazy-load the CodeMirror editor so it lands in an on-demand chunk
- Split App and DatabaseForm into smaller components and hooks
- Add doctor.config.json suppressing documented false positives
- Add a React Doctor GitHub Action that scans every PR
