+++ .github/workflows/react-doctor.yml
@@ -1,0 +1,32 @@
+name: React Doctor
+
+on:
+  # Scans the PR's changed files and posts a sticky summary comment listing
+  # only the new issues introduced relative to the merge base.
+  pull_request:
+    types: [opened, synchronize, reopened, ready_for_review]
+  # Scans main on every push to track the health-score trend.
+  push:
+    branches: [main]
+
+permissions:
+  contents: read
+  pull-requests: write
+  issues: write
+  statuses: write
+
+# Cancels any in-flight scan for the same PR (or branch, on push) the moment
+# a new commit arrives, so reviewers only ever see the latest run.
+concurrency:
+  group: react-doctor-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  react-doctor:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v5
+        with:
+          fetch-depth: 0
+
+      - uses: millionco/react-doctor@v2
