+++ package.json
@@ -23,16 +23,10 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@ai-sdk/anthropic": "^2.0.45",
-    "@ai-sdk/google": "^2.0.40",
-    "@ai-sdk/openai": "^2.0.71",
-    "@ai-sdk/react": "^2.0.98",
     "@babel/runtime": "^7.28.4",
     "@codemirror/autocomplete": "^6.20.0",
     "@codemirror/lang-sql": "^6.10.0",
     "@codemirror/language": "^6.11.3",
-    "@codemirror/lint": "^6.9.2",
-    "@codemirror/search": "^6.5.11",
     "@codemirror/state": "^6.5.2",
     "@codemirror/theme-one-dark": "^6.1.3",
     "@codemirror/view": "^6.38.8",
@@ -45,7 +39,6 @@
     "@lezer/highlight": "^1.2.3",
     "@libsql/client": "^0.15.15",
     "@radix-ui/react-context-menu": "^2.2.16",
-    "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-label": "^2.1.8",
     "@radix-ui/react-select": "^2.2.6",
     "@radix-ui/react-separator": "^1.1.8",
@@ -69,14 +62,12 @@
     "hono": "^4.10.6",
     "lucide-react": "^0.554.0",
     "mysql2": "^3.15.3",
-    "next-themes": "^0.4.6",
     "pg": "^8.16.3",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
     "react-hook-form": "^7.66.1",
     "react-hotkeys-hook": "^5.2.1",
     "react-redux": "^9.2.0",
-    "react-router-dom": "^7.9.6",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.4.0",
     "tiny-invariant": "^1.3.3",
