+++ src/app/components/ThemeProvider.tsx
@@ -1,4 +1,4 @@
-import { createContext, use, type ReactNode } from 'react'
+import { createContext, type ReactNode } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
 
 import { useTheme } from '@/app/hooks/useTheme'
@@ -16,14 +16,4 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   })
 
   return <ThemeContext.Provider value={theme}>{children}</ThemeContext.Provider>
-}
-
-export function useThemeContext() {
-  const context = use(ThemeContext)
-
-  if (!context) {
-    throw new Error('useThemeContext must be used within ThemeProvider')
-  }
-
-  return context
 }
