+++ src/app/components/ui/button.tsx
@@ -54,4 +54,4 @@ function Button({
   )
 }
 
-export { Button, buttonVariants }
+export { Button }
