--- src/app/components/ui/skeleton.tsx
@@ -1,13 +1,0 @@
-import { cn } from '@/app/lib/utils'
-
-function Skeleton({ className, ...props }: React.ComponentProps<'div'>) {
-  return (
-    <div
-      data-slot="skeleton"
-      className={cn('bg-surface-2 animate-pulse rounded-md', className)}
-      {...props}
-    />
-  )
-}
-
-export { Skeleton }
